split func to proper files
