+++ pocket.go
@@ -5,82 +5,24 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/pkg/errors"
 	"io/ioutil"
 	"net/http"
 	"net/url"
-	"strings"
 	"time"
+
+	"github.com/pkg/errors"
 )
 
 const (
-	host         = "https://getpocket.com/v3"
-	authorizeUrl = "https://getpocket.com/auth/authorize?request_token=%s&redirect_uri=%s"
+	host = "https://getpocket.com/v3"
 
-	endpointAdd          = "/add"
-	endpointRequestToken = "/oauth/request"
-	endpointAuthorize    = "/oauth/authorize"
+	endpointAdd = "/add"
 
 	// xErrorHeader used to parse error message from Headers on non-2XX responses
 	xErrorHeader = "X-Error"
 
 	defaultTimeout = 5 * time.Second
 )
-
-type (
-	requestTokenRequest struct {
-		ConsumerKey string `json:"consumer_key"`
-		RedirectURI string `json:"redirect_uri"`
-	}
-
-	authorizeRequest struct {
-		ConsumerKey string `json:"consumer_key"`
-		Code        string `json:"code"`
-	}
-
-	AuthorizeResponse struct {
-		AccessToken string `json:"access_token"`
-		Username    string `json:"username"`
-	}
-
-	addRequest struct {
-		URL         string `json:"url"`
-		Title       string `json:"title,omitempty"`
-		Tags        string `json:"tags,omitempty"`
-		AccessToken string `json:"access_token"`
-		ConsumerKey string `json:"consumer_key"`
-	}
-
-	// AddInput holds data necessary to create new item in Pocket list
-	AddInput struct {
-		URL         string
-		Title       string
-		Tags        []string
-		AccessToken string
-	}
-)
-
-func (i AddInput) validate() error {
-	if i.URL == "" {
-		return errors.New("required URL values is empty")
-	}
-
-	if i.AccessToken == "" {
-		return errors.New("access token is empty")
-	}
-
-	return nil
-}
-
-func (i AddInput) generateRequest(consumerKey string) addRequest {
-	return addRequest{
-		URL:         i.URL,
-		Tags:        strings.Join(i.Tags, ","),
-		Title:       i.Title,
-		AccessToken: i.AccessToken,
-		ConsumerKey: consumerKey,
-	}
-}
 
 // Client is a getpocket API client
 type Client struct {
@@ -102,74 +44,7 @@ func NewClient(consumerKey string) (*Client, error) {
 	}, nil
 }
 
-// GetRequestToken obtains the request token that is used to authorize user in your application
-func (c *Client) GetRequestToken(ctx context.Context, redirectUrl string) (string, error) {
-	inp := &requestTokenRequest{
-		ConsumerKey: c.consumerKey,
-		RedirectURI: redirectUrl,
-	}
-
-	values, err := c.doHTTP(ctx, endpointRequestToken, inp)
-	if err != nil {
-		return "", err
-	}
-
-	if values.Get("code") == "" {
-		return "", errors.New("empty request token in API response")
-	}
-
-	return values.Get("code"), nil
-}
-
-// GetAuthorizationURL generates link to authorize user
-func (c *Client) GetAuthorizationURL(requestToken, redirectUrl string) (string, error) {
-	if requestToken == "" || redirectUrl == "" {
-		return "", errors.New("empty params")
-	}
-
-	return fmt.Sprintf(authorizeUrl, requestToken, redirectUrl), nil
-}
-
-// Authorize generates access token for user, that authorized in your app via link
-func (c *Client) Authorize(ctx context.Context, requestToken string) (*AuthorizeResponse, error) {
-	if requestToken == "" {
-		return nil, errors.New("empty request token")
-	}
-
-	inp := &authorizeRequest{
-		Code:        requestToken,
-		ConsumerKey: c.consumerKey,
-	}
-
-	values, err := c.doHTTP(ctx, endpointAuthorize, inp)
-	if err != nil {
-		return nil, err
-	}
-
-	accessToken, username := values.Get("access_token"), values.Get("username")
-	if accessToken == "" {
-		return nil, errors.New("empty access token in API response")
-	}
-
-	return &AuthorizeResponse{
-		AccessToken: accessToken,
-		Username:    username,
-	}, nil
-}
-
-// Add creates new item in Pocket list
-func (c *Client) Add(ctx context.Context, input AddInput) error {
-	if err := input.validate(); err != nil {
-		return err
-	}
-
-	req := input.generateRequest(c.consumerKey)
-	_, err := c.doHTTP(ctx, endpointAdd, req)
-
-	return err
-}
-
-func (c *Client) doHTTP(ctx context.Context, endpoint string, body interface{}) (url.Values, error) {
+func (c *Client) DoHTTP(ctx context.Context, endpoint string, body interface{}) (url.Values, error) {
 	b, err := json.Marshal(body)
 	if err != nil {
 		return url.Values{}, errors.WithMessage(err, "failed to marshal input body")
